src/requirements.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
cycler==0.10.0
dataclasses==0.8
deap==1.3.1
decorator==4.4.2
dgl-cu111==0.7.2
easydict==1.9
flatbuffers==1.12
gast==0.3.3
google-auth==1.31.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
graphviz==0.16
grpcio==1.32.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.5.0
joblib==1.1.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
littleutils==0.2.2
Markdown==3.3.4
matplotlib==3.0.3
networkx==2.5.1
nibabel==3.2.1
nilearn==0.8.1
numpy==1.19.5
oauthlib==3.1.1
ogb-lite==0.0.3
opt-einsum==3.3.0
outdated==0.2.1
packaging==21.3
pandas==1.1.5
Pillow==8.1.2
prettytable==2.1.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pydotplus==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
sklearn-genetic==0.4.1
tensorboard==2.4.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-addons==0.14.0
tensorflow-determinism==0.3.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
threadpoolctl==3.0.0
tqdm==4.60.0
typeguard==2.13.3
typing-extensions==3.7.4.3
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
xlrd==2.0.1
zh==1.0.0
zipp==3.4.1
